+++ packages/checkout/src/contexts/index.ts
@@ -1,2 +1,3 @@
+export * from './AddFundsModal'
 export * from './CheckoutModal'
 export * from './Navigation'
