+++ packages/checkout/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useAddFundsModal'
 export * from './useCheckoutModal'
 export * from './useNavigation'
 export * from './useModalTheme'
