+++ packages/checkout/src/hooks/useAddFundsModal.ts
@@ -1,0 +1,7 @@
+import { useAddFundsModalContext } from '../contexts/AddFundsModal'
+
+export const useAddFundsModal = () => {
+  const { triggerAddFunds, closeAddFunds, addFundsSettings } = useAddFundsModalContext()
+
+  return { triggerAddFunds, closeAddFunds, addFundsSettings }
+}
