+++ packages/checkout/src/views/index.ts
@@ -2,3 +2,4 @@ export * from './PendingTransaction'
 export * from './TransactionSuccess'
 export * from './TransactionError'
 export * from './CheckoutSelection'
+export * from './AddFunds'
